refactor(TAT-138): pull in changes from homepage
