+++ src/views/CalculatorPage.vue
