+++ src/views/TopTenResults.vue
@@ -2,7 +2,7 @@
     <div id="calculator">
         <div class="w-5/6 mx-auto pt-20">
             <h1 class="text-center uppercase font-bold text-5xl">Your Top 10 Techniques</h1>
-            <div class="w-full h-auto lg:border-0 border-[1px] border-mitre-black mt-4 mb-6 py-2 lg:py-0">
+            <div class="w-full h-auto lg:border-0 border-[1px] border-ctid-black mt-4 mb-6 py-2 lg:py-0">
                 <SystemScoreSection />
             </div>
         </div>
@@ -15,7 +15,7 @@
                 <div class="calculator-details">
                     <div class="container-header">
                         <h2>
-                            <span class=" text-mitre-light-purple">
+                            <span class=" text-ctid-light-purple">
                                 {{ topTen[activeItemId].tid }}
                             </span>
                             {{ topTen[activeItemId].name }}
@@ -38,9 +38,9 @@
 <script lang="ts">
 import { defineComponent } from "vue";
 import { useCalculatorStore } from "../stores/calculator.store";
-import TopTenSidebar from "./TopTenSidebar.vue"
-import TopTenDetails from "./TopTenDetails.vue"
-import TopTenAccordion from "./TopTenAccordion.vue"
+import TopTenSidebar from "../components/TopTenSidebar.vue"
+import TopTenDetails from "../components/TopTenDetails.vue"
+import TopTenAccordion from "../components/TopTenAccordion.vue"
 import SystemScoreSection from "../components/SystemScoreSection.vue"
 
 export default defineComponent({
