+++ tailwind.config.js
@@ -5,24 +5,14 @@ module.exports = {
   theme: {
     extend: {
       colors: {
-        mitre: {
-          blue: "#005B94",
-          highlighter: "#FFF601",
-          "dark-navy": "#0B2338",
-          navy: "#0D2F4F",
-          "light-blue": "#87DEFF",
-          black: "#111921",
-          "dark-gray": "#7E8284",
-          silver: "#D4D4D3",
-          "light-silver": "#F1F3F4",
-          "primary-purple": "#7450B2",
-          "light-purple": "#CEBEEA",
-        },
         ctid: {
           "primary-purple": "#5C6FC6",
           navy: "#212C5E",
           black: "#0B2338",
-          "light-purple": "#BCA9DD",
+          "primary-purple": "#7450B2",
+          "light-purple": "#CEBEEA",
+          blue: "#005B94",
+          "dark-navy": "#0B2338",
         },
       },
     },
